+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 2019-11-13
+### Changed
+* Added `--list-arn-types` under the query command.
+  - This allows the user to query available ARN types and RAW ARN pairs
+  - Added test case for the new query command
 ## 2019-11-08
 ### Added
 * Query capability to assist users in identifying actions to supply in future #16 feature